+++ chrolo.personal_functions.lua
@@ -5,8 +5,10 @@
 script_name = "Chrolo's Scripts"
 script_description = "Chrolo's personal scripts and macros."
 script_author = "Chrolo"
-script_version = "v0.1"
+script_version = "0.1.0"
+script_namespace="chrolo.personal_functions"
 
+---------------------------------------------------------------------------------------------------------------------
 function backup_lines(subs, selection)
 	local flag = true
 	local j = 0
@@ -44,6 +46,7 @@ end
 --Macro registration
 aegisub.register_macro("Chrolo/Line Backup", "Duplicates and Comments the selected lines.", backup_lines)
 
+---------------------------------------------------------------------------------------------------------------------
 function sel_between_comments(subs,selection)
 	local last_line=0
 	local last_ind=0
@@ -68,56 +71,76 @@ function sel_between_comments(subs,selection)
 end
 aegisub.register_macro("Chrolo/Select to next comment", "Add lines to selection until it finds a comment",sel_between_comments)
 
+---------------------------------------------------------------------------------------------------------------------
+
+function copy_timing_from_above(subs,selection)
+	
+	local prev_line= selection[1]-1
+
+	for _, i in ipairs(selection) do
+		line = subs[i]
+		line.start_time = subs[prev_line].start_time
+		line.end_time = subs[prev_line].end_time
+		subs[i] = line
+	end
+	
+	return selection
+end
+aegisub.register_macro("Chrolo/Copy Timing", "Copy the timing of the line above selection into the selection",copy_timing_from_above)
+
+
+
+
+
+
+
+
+
+
+
+
+
+------------------------------------------------------------
+--	Test Area											  --
+------------------------------------------------------------
+
+-------------
 function test(subs, sel)
 	
-	libLoad = require'chroloKfxLib'
-	local libChroloKfx , eff_lib = libLoad()
-	
-	if not libChroloKfx then
-		aegisub.debug.out("Lib Chrolo Failed to load")
-		return false
-	end
-	if not eff_lib then
-		aegisub.debug.out("Effect Library Failed to load")
-		return false
-	end
-	util = require 'aegisub.util'
+	--util = require 'aegisub.util'
 	require 'print_r'
+	require 'karaskel'
 	
-	local bpm=175
+
+	
+---------------------------------------------------------------------
+
 	---[[
-	--aegisub.debug.out(str)
-	local line = subs[#subs]
-	line.layer=1
-	line.start_time=line.end_time
-	line.end_time=line.end_time+5000
-	eff_lib.Shuffle:setShuffle(100,100)
-	eff_lib.Shuffle:setShuffleFade(0.5)
-	str = libChroloKfx.KFX_line_out(eff_lib.RainbowText,libChroloKfx.bpm_to_ms(5000),5000,(line.start_time%libChroloKfx.bpm_to_ms(bpm)))
-	line.text="{"..str.."}Test Text"
-	subs.append(line)
-	
-	
-	line.start_time=line.end_time
-	--entrance line
-	local time_in=1000
-	line.layer=0
-	line.end_time=line.start_time+time_in
-	eff_lib.Twirl_In:setIntime(time_in)
-	str = libChroloKfx.KFX_line_out(eff_lib.Twirl_In)
-	line.text="{"..str.."}Test Text2"
-	subs.append(line)
-	
-	--next line
-	line.start_time=line.end_time
-	line.end_time=line.end_time+5000
-	eff_lib.RainbowText:setColourTag(3)
-	str = libChroloKfx.KFX_line_out(eff_lib.LRBounce,libChroloKfx.bpm_to_ms(bpm),5000,(line.start_time%libChroloKfx.bpm_to_ms(bpm))) .. libChroloKfx.KFX_line_out(eff_lib.RainbowText,libChroloKfx.bpm_to_ms(5000),5000,(line.start_time%libChroloKfx.bpm_to_ms(bpm)))
-	line.text="{\\pos()"..str.."}Test Text2"
-	subs.append(line)
+	chroloLibLoad = require 'chrolo.lib'
+	chLib = chroloLibLoad("debug_level",1)
+	if chLib == nil then
+		aegisub.debug.out("Chrolo lib not found\n")
+		return sel
+	end
 	--]]
 	
-	
+	--[[
+	for i = 1, #subs do
+		aegisub.debug.out(string.format("%s\n",print_r(subs[i],"Line")))
+	end
+	--]]
+	---[[
+	for _,i in ipairs(sel) do
+		bounds = chLib.getTextBoundCoords(subs, subs[i])
+		--bounds = {chLib.getTextBound(subs, subs[i])}
+		
+		aegisub.debug.out(string.format("%s\n",print_r(bounds,"Coords")))
+		--aegisub.debug.out(string.format("\n{\\p1}m 0 0 l %g 0 %g %g 0 %g\n",bounds[1],bounds[1],bounds[2], bounds[2]))
+		aegisub.debug.out(string.format("\n{\\p1\\an7\\pos(0,0)}m %g %g  l %g %g %g %g %g %g\n",bounds[1][1],bounds[1][2],bounds[2][1],bounds[2][2],bounds[3][1],bounds[3][2],bounds[4][1],bounds[4][2]))
+	end
+	--]]
+
+-----------------------------------------------------------------------
 	return sel
 end