+++ changelog.md
@@ -1,13 +1,36 @@
 ##Changelog
 
-**13/10/2015**
- - added blank returns to all effect_lib param functions, so they can be used in kfx codeblocks
- -
+**11/01/2016**
+- New "chrolo.lib" file
+  - Has various functions to determine size / co ordinates of text
+  - will be update with more....
+  
+- New macro: Change Alignment
+  - Aims to change the alignement of text without modifying it's position
+  - cannot handle lines without '\pos' set (due to limitations in my library)
+
+**13/10/2015**  
+- added blank returns to all effect_lib param functions, so they can be used in kfx codeblocks
  
  
- ##To do
- [ ] Change how shuffle text's extra data is prepended
- [ ] actually do kfx for a show to test my stuff
- [ ] add millions more effects to the effect library
- [ ] get "QuickKFX" up to a working standard (soon™)
- [ ] get used to github markdown language
+##To do
+- [ ] Get fancy and put dep control into my stuff
+
+###Alignment Macro
+- [ ] have macro change \move tags appropriately aswell.
+
+###Lib stuff
+- [ ] Improve chrolo.lib's TextBound function to handle: \shad, \fax, \fry, \frx... etc
+- [ ] Add ability to determine default line position when \pos not specified.
+
+###KFX lib stuff
+- [ ] Change how shuffle text's extra data is prepended
+
+- [ ] actually do kfx for a show to test my stuff
+
+- [ ] add millions more effects to the effect library
+
+- [ ] get "QuickKFX" up to a working standard (soon™)
+
+###Misc
+- [ ] get used to github markdown language